chore: setup deno
