+++ deno.json
@@ -1,0 +1,6 @@
+{
+  "compilerOptions": {
+    "declaration": true,
+    "sourceMap": true
+  }
+}
