+++ src/parsers/any-parser.ts
@@ -1,4 +1,4 @@
-import { IParser, OnParse } from "./interfaces";
+import { IParser, OnParse } from "./interfaces.ts";
 
 export class AnyParser implements IParser<unknown, any> {
   constructor(
