+++ src/parsers/bool-parser.ts
@@ -1,5 +1,5 @@
-import { isNumber } from "./index";
-import { IParser, OnParse } from "./interfaces";
+import { isNumber } from "./index.ts";
+import { IParser, OnParse } from "./interfaces.ts";
 
 export class BoolParser implements IParser<unknown, boolean> {
   constructor(
