+++ src/tests.ts
@@ -1,0 +1,2 @@
+import "./matches.test.ts";
+// import "./transformers/typescript-types.test.ts";
